Hooked up Location and Radius to UiState

+++ core/data/src/main/java/com/ethereal/data/repository/OfflineFirstUserDataRepository.kt
@@ -24,7 +24,7 @@ class OfflineFirstUserDataRepository @Inject constructor(
         }
     }
 
-    override suspend fun setDefaultRadius(radius: Double) {
+    override suspend fun setDefaultRadius(radius: Int) {
         CoroutineScope(ioDispatcher).launch {
             bdUserPreferences.setDefaultRadius(radius)
         }
