+++ core/model/src/main/java/com/ethereal/model/data/DateDetails.kt
@@ -15,5 +15,5 @@ data class MapData(
     val mapLoading: Boolean = true,
     val userLocation: GeoPoint?,
     val cityState: String = "",
-    val radiusMiles: Double
+    val radiusMiles: Int
 )
