+++ core/model/src/main/java/com/ethereal/model/data/UserData.kt
@@ -2,5 +2,5 @@ package com.ethereal.model.data
 
 data class UserData(
     val authenticationToken: String,
-    val defaultRadius: Double
+    val defaultRadius: Int
 )
