+++ core/model/src/main/java/com/ethereal/model/utils/GeoConversions.kt
@@ -5,7 +5,7 @@ import com.google.android.gms.maps.model.LatLng
 
 
 private const val METERS_PER_MILE = 1609.344f
-fun milesToMeters(mi: Double): Double = mi * METERS_PER_MILE
+fun milesToMeters(mi: Int): Double = (mi * METERS_PER_MILE).toDouble()
 fun metersToMiles(m: Double): Double = m / METERS_PER_MILE
 
 fun GeoPoint.toLatLng(): LatLng = LatLng(lat, lng)
